Adds a snapshot unit test for getConnections and rearranges some code to make testing easier
